+++ modules/options_custom_items.lua
@@ -6,6 +6,7 @@ local L = LibStub("AceLocale-3.0"):GetLocale("EPGP")
 local LIU = LibStub("LibItemUtils-1.0")
 local LN = LibStub("LibLocalConstant-1.0")
 local LUI = LibStub("LibEPGPUI-1.0")
+local LOor = LibStub("LibEpgpOorProfile-1.0")
 
 local LOCAL_NAME = LN:LocalName()
 
@@ -381,7 +382,7 @@ local function ItemRemoveButtonOnClickFunc(self)
 end
 
 local function CreateAddFrame(parent)
-  addFrame = CreateFrame("Frame", nil, parent)
+  addFrame = CreateFrame("Frame", nil, parent, BackdropTemplateMixin and "BackdropTemplate");
   addFrame:SetPoint("TOPLEFT")
   addFrame:SetPoint("RIGHT")
 
@@ -391,7 +392,7 @@ local function CreateAddFrame(parent)
   iconF:SetPoint("TOPLEFT")
   addFrame.iconF = iconF
 
-  local iconFrame = CreateFrame("Frame", nil, addFrame)
+  local iconFrame = CreateFrame("Frame", nil, addFrame, BackdropTemplateMixin and "BackdropTemplate");
   iconFrame:ClearAllPoints()
   iconFrame:SetAllPoints(iconF)
   iconFrame:SetScript("OnEnter", ItemIconFrameOnEnterFunc)
@@ -511,7 +512,7 @@ local function AddTitle(f, name, width, top, left)
 end
 
 local function AddOneItemFrame(parent, top)
-  local f = CreateFrame("Frame", nil, parent)
+  local f = CreateFrame("Frame", nil, parent, BackdropTemplateMixin and "BackdropTemplate");
   f:SetPoint("LEFT")
   f:SetPoint("RIGHT")
   if top then
@@ -526,7 +527,7 @@ local function AddOneItemFrame(parent, top)
   iconF:SetPoint("TOPLEFT")
   f.iconF = iconF
 
-  local iconFrame = CreateFrame("Frame", nil, f)
+  local iconFrame = CreateFrame("Frame", nil, f, BackdropTemplateMixin and "BackdropTemplate");
   iconFrame:ClearAllPoints()
   iconFrame:SetAllPoints(iconF)
   iconFrame:SetScript("OnEnter", ItemIconFrameOnEnterFunc)
@@ -655,7 +656,7 @@ function mod:FillFrame(f, parent)
   t = AddTitle(f, "GP1", columnWidth.gp, addFrame, t)
   t = AddTitle(f, "GP2", columnWidth.gp, addFrame, t)
 
-  containerFrame = CreateFrame("Frame", nil, f)
+  containerFrame = CreateFrame("Frame", nil, f, BackdropTemplateMixin and "BackdropTemplate");
   containerFrame:SetPoint("TOP", t, "BOTTOM")
   containerFrame:SetPoint("LEFT")
   containerFrame:SetWidth(columnWidthTotal + 27)
@@ -741,7 +742,7 @@ local function AddDefaultData()
   local ci = vars.customItems
   local faction = UnitFactionGroup("player")
   for i, v in pairs(CUSTOM_ITEM_DATA) do
-    if not v[5] or v[5] == faction then
+    if not v[4] or v[4] == faction then
       if not ci[i] then
         local equipLocKey = EQUIPLOC_INDEX[v[3]] or EQUIPLOC_CUSTOM_SCALE_INDEX
         ci[i] = {
@@ -753,6 +754,14 @@ local function AddDefaultData()
           gp1 = v[6],
           gp2 = v[7],
         }
+
+        if equipLocKey == EQUIPLOC_CUSTOM_SCALE_INDEX then
+          if v[5] then ci[i].s1 = v[5] end
+          if v[6] then ci[i].s2 = v[6] end
+        elseif equipLocKey == EQUIPLOC_CUSTOM_GP_INDEX then
+          if v[5] then ci[i].gp1 = v[5] end
+          if v[6] then ci[i].gp2 = v[6] end
+        end
       end
     else
       ci[i] = nil
@@ -781,4 +790,15 @@ function mod:OnEnable()
 end
 
 function mod:OnDisable()
+end
+
+function mod:CheckGuildConfig(guild, realm)
+  if (guild == "Order Of Rhonin" or guild == "EPGP test") and realm == "艾隆纳亚" then
+    if EPGP.db.profile.customItems then
+      table.wipe(EPGP.db.profile.customItems)
+    else
+      EPGP.db.profile.customItems = {}
+    end
+    CUSTOM_ITEM_DATA = LOor:GetCustomItemsProfile()
+  end
 end