+++ modules/ui.lua
@@ -2,6 +2,7 @@
 
 local mod = EPGP:NewModule("ui")
 local L = LibStub("AceLocale-3.0"):GetLocale("EPGP")
+local LE = LibStub("AceLocale-3.0"):GetLocale("LibEncounters")
 local GS = LibStub("LibGuildStorage-1.2")
 local GP = LibStub("LibGearPoints-1.3")
 local DLG = LibStub("LibDialog-1.0")
@@ -96,7 +97,7 @@ end
 
 local function CreateEPGPFrame()
   -- EPGPFrame
-  local f = CreateFrame("Frame", "EPGPFrame", UIParent)
+  local f = CreateFrame("Frame", "EPGPFrame", UIParent, BackdropTemplateMixin and "BackdropTemplate");
   f:Hide()
   f:SetToplevel(true)
   f:EnableMouse(true)
@@ -161,7 +162,7 @@ local function CreateEPGPFrame()
 end
 
 local function CreateEPGPExportImportFrame()
-  local f = CreateFrame("Frame", "EPGPExportImportFrame", UIParent)
+  local f = CreateFrame("Frame", "EPGPExportImportFrame", UIParent, BackdropTemplateMixin and "BackdropTemplate");
   f:Hide()
   f:SetPoint("CENTER")
   f:SetFrameStrata("TOOLTIP")
@@ -228,7 +229,7 @@ local function CreateEPGPExportImportFrame()
         self.editbox:SetScript("OnTextChanged", nil)
       elseif self.exportDetail then
         self.help:SetText(L["EXPORT_DETAIL_DESC"] .. "\n" ..
-                          "https://epgp.dj173.com - " .. _G.ZHCN .. " - " .. L["Web & WeChat Mini Program"])
+                          "https://www.wowepgp.com - " .. _G.ZHCN .. " - " .. L["Web & WeChat Mini Program"])
         self.button1:Show()
         self.button1:SetText(CLOSE)
         self.button1:SetPoint("CENTER", self, "CENTER")
@@ -365,7 +366,7 @@ local function CreateTable(parent, texts, widths, justfiesH, rightPadding)
   end
 
   -- Make a frame for the rows
-  local rowFrame = CreateFrame("Frame", nil, parent)
+  local rowFrame = CreateFrame("Frame", nil, parent, BackdropTemplateMixin and "BackdropTemplate");
   rowFrame:SetPoint("TOP", parent.headers[#parent.headers], "BOTTOM")
   rowFrame:SetPoint("BOTTOMLEFT")
   rowFrame:SetPoint("BOTTOMRIGHT", parent, "BOTTOMRIGHT", -rightPadding, 0)
@@ -391,7 +392,7 @@ local function CreateTable(parent, texts, widths, justfiesH, rightPadding)
 end
 
 local function CreateEPGPLogFrame()
-  local f = CreateFrame("Frame", "EPGPLogFrame", EPGPFrame)
+  local f = CreateFrame("Frame", "EPGPLogFrame", EPGPFrame, BackdropTemplateMixin and "BackdropTemplate");
   table.insert(SIDEFRAMES, f)
 
   f:SetResizable(true)
@@ -544,7 +545,7 @@ local function CreateEPGPLogFrame()
       end
     end)
 
-  local scrollParent = CreateFrame("Frame", nil, f)
+  local scrollParent = CreateFrame("Frame", nil, f, BackdropTemplateMixin and "BackdropTemplate");
   scrollParent:SetPoint("TOP", t, "TOP", 0, -16)
   scrollParent:SetPoint("BOTTOM", redo, "TOP", 0, 0)
   scrollParent:SetPoint("LEFT", f, "LEFT", 16, 0)
@@ -918,13 +919,14 @@ local function EPGPSideFrameEPDropDown_SetList(dropDown)
   --   end
   -- end
 
-  tinsert(list, _G.DUNGEON_FLOOR_MOLTENCORE1)
-  tinsert(list, _G.DUNGEON_FLOOR_ONYXIASLAIR1)
-  tinsert(list, L["Blackwing Lair"])
-  tinsert(list, _G.DUNGEON_FLOOR_ZULGURUB1)
-  tinsert(list, _G.DUNGEON_FLOOR_RUINSOFAHNQIRAJ1)
-  tinsert(list, L["Temple of Ahn'Qiraj"])
-  tinsert(list, L["Naxxramas"])
+  tinsert(list, LE["Karazhan"])
+  tinsert(list, LE["Magtheridon's Lair"])
+  tinsert(list, LE["Gruul's Lair"])
+  tinsert(list, LE["Coilfang: Serpentshrine Cavern"])
+  tinsert(list, LE["Tempest Keep"])
+  tinsert(list, LE["The Battle for Mount Hyjal"])
+  tinsert(list, LE["Black Temple"])
+  tinsert(list, LE["The Sunwell"])
   tinsert(list, OTHER)
 
   dropDown:SetList(list)
@@ -1582,7 +1584,7 @@ local function AddDecayControls(frame)
 end
 
 local function CreateEPGPSideFrame(self)
-  local f = CreateFrame("Frame", "EPGPSideFrame", EPGPFrame)
+  local f = CreateFrame("Frame", "EPGPSideFrame", EPGPFrame, BackdropTemplateMixin and "BackdropTemplate");
   table.insert(SIDEFRAMES, f)
 
   f:Hide()
@@ -1618,11 +1620,11 @@ local function CreateEPGPSideFrame(self)
   local cb = CreateFrame("Button", nil, f, "UIPanelCloseButton")
   cb:SetPoint("TOPRIGHT", f, "TOPRIGHT", -2, -3)
 
-  local gpFrame = CreateFrame("Frame", nil, f)
+  local gpFrame = CreateFrame("Frame", nil, f, BackdropTemplateMixin and "BackdropTemplate");
   gpFrame:SetPoint("TOPLEFT", f, "TOPLEFT", 15, -30)
   gpFrame:SetPoint("TOPRIGHT", f, "TOPRIGHT", -15, -30)
 
-  local epFrame = CreateFrame("Frame", nil, f)
+  local epFrame = CreateFrame("Frame", nil, f, BackdropTemplateMixin and "BackdropTemplate");
   epFrame:SetPoint("TOPLEFT", gpFrame, "BOTTOMLEFT", 0, -15)
   epFrame:SetPoint("TOPRIGHT", gpFrame, "BOTTOMRIGHT", -15, -15)
 
@@ -1662,7 +1664,7 @@ local function CreateEPGPSideFrame(self)
 end
 
 local function CreateEPGPSideFrame2()
-  local f = CreateFrame("Frame", "EPGPSideFrame2", EPGPFrame)
+  local f = CreateFrame("Frame", "EPGPSideFrame2", EPGPFrame, BackdropTemplateMixin and "BackdropTemplate");
   table.insert(SIDEFRAMES, f)
 
   f:Hide()
@@ -1682,15 +1684,15 @@ local function CreateEPGPSideFrame2()
   local cb = CreateFrame("Button", nil, f, "UIPanelCloseButton")
   cb:SetPoint("TOPRIGHT", f, "TOPRIGHT", -2, -3)
 
-  local epFrame = CreateFrame("Frame", nil, f)
+  local epFrame = CreateFrame("Frame", nil, f, BackdropTemplateMixin and "BackdropTemplate");
   epFrame:SetPoint("TOPLEFT", f, "TOPLEFT", 15, -15)
   epFrame:SetPoint("TOPRIGHT", f, "TOPRIGHT", -30, -15)
 
-  local customFrame = CreateFrame("Frame", nil, f)
+  local customFrame = CreateFrame("Frame", nil, f, BackdropTemplateMixin and "BackdropTemplate");
   customFrame:SetPoint("TOPLEFT", epFrame, "BOTTOMLEFT", 0, -15)
   customFrame:SetPoint("TOPRIGHT", epFrame, "BOTTOMRIGHT", 0, -15)
 
-  local decayFrame = CreateFrame("Frame", nil, f)
+  local decayFrame = CreateFrame("Frame", nil, f, BackdropTemplateMixin and "BackdropTemplate");
   decayFrame:SetPoint("TOPLEFT", customFrame, "BOTTOMLEFT", 0, -15)
   decayFrame:SetPoint("TOPRIGHT", customFrame, "BOTTOMRIGHT", 0, -15)
 
@@ -1743,7 +1745,7 @@ local function CreateEPGPSideFrame2()
 end
 
 local function CreateEPGPLootFrame()
-  local f = CreateFrame("Frame", "EPGPLootFrame", EPGPFrame)
+  local f = CreateFrame("Frame", "EPGPLootFrame", EPGPFrame, BackdropTemplateMixin and "BackdropTemplate");
   table.insert(SIDEFRAMES, f)
 
   f:Hide()
@@ -1767,7 +1769,7 @@ local function CreateEPGPLootFrame()
   local cb = CreateFrame("Button", nil, f, "UIPanelCloseButton")
   cb:SetPoint("TOPRIGHT", f, "TOPRIGHT", -2, -3)
 
-  local itemFrame = CreateFrame("Frame", nil, f)
+  local itemFrame = CreateFrame("Frame", nil, f, BackdropTemplateMixin and "BackdropTemplate");
   itemFrame:SetPoint("TOPLEFT", f, "TOPLEFT", 15, -15)
   itemFrame:SetPoint("TOPRIGHT", f, "TOPRIGHT", -15, -15)
   itemFrame:SetScript("OnShow",
@@ -1782,7 +1784,7 @@ local function CreateEPGPLootFrame()
 end
 
 local function CreateEPGPOptionsFrame()
-  local f = CreateFrame("Frame", "EPGPOptionsFrame", EPGPFrame)
+  local f = CreateFrame("Frame", "EPGPOptionsFrame", EPGPFrame, BackdropTemplateMixin and "BackdropTemplate");
   table.insert(SIDEFRAMES, f)
 
   f:Hide()
@@ -1820,7 +1822,7 @@ end
 
 local function CreateEPGPFrameStandings()
   -- Make the show everyone checkbox
-  local f = CreateFrame("Frame", nil, EPGPFrame)
+  local f = CreateFrame("Frame", nil, EPGPFrame, BackdropTemplateMixin and "BackdropTemplate");
   f:SetHeight(28)
   f:SetPoint("TOPRIGHT", EPGPFrame, "TOPRIGHT", -42, -38)
 
@@ -1884,7 +1886,7 @@ local function CreateEPGPFrameStandings()
   CreateEPGPOptionsFrame()
 
   -- Make the main frame
-  local main = CreateFrame("Frame", nil, EPGPFrame)
+  local main = CreateFrame("Frame", nil, EPGPFrame, BackdropTemplateMixin and "BackdropTemplate");
   main:SetWidth(325)
   main:SetHeight(358)
   main:SetPoint("TOPLEFT", EPGPFrame, 19, -72)
@@ -2010,7 +2012,7 @@ local function CreateEPGPFrameStandings()
   end
 
   -- Make the table frame
-  local tabl = CreateFrame("Frame", nil, main)
+  local tabl = CreateFrame("Frame", nil, main, BackdropTemplateMixin and "BackdropTemplate");
   tabl:SetPoint("TOPLEFT")
   tabl:SetPoint("TOPRIGHT")
   tabl:SetPoint("BOTTOM", modeText, "TOP")
@@ -2121,7 +2123,7 @@ local function CreateEPGPFrameStandings()
         local name = EPGP:GetMember(j)
         row.name = name
         row.cells[1]:SetText(Ambiguate(row.name, "short"))
-        local c = RAID_CLASS_COLORS[EPGP:GetClass(row.name)]
+        local c = _G.RAID_CLASS_COLORS[EPGP:GetClass(row.name)]
         row.cells[1]:SetTextColor(c.r, c.g, c.b)
         local ep, gp = EPGP:GetEPGP(row.name)
         row.cells[2]:SetText(ep)